src/main/java/MoodAnalyser.java
public class MoodAnalyser {
    private String message;

    public MoodAnalyser() {
    }

    public MoodAnalyser(String message) {
        this.message = message;
    }

    public String analyseMood(String message) throws MoodAnalysisException {
        this.message = message;
        try {
            if (message.length()==0)
                throw new MoodAnalysisException(MoodAnalysisException.ExceptionType.ENTERED_EMPTY,"Please Enter Correct Message");
            if (message.contains("sad"))
                return "sad";
            return "happy";
        }catch (NullPointerException result){
            throw new MoodAnalysisException(MoodAnalysisException.ExceptionType.ENTERED_NULL,"Please Enter Correct Message");
        }
    }

    public String analyseMood() {
        return "sad";
    }
}